requirements.txt
aiohttp
allure-behave
allure-pytest
async-timeout
pytest
ujson
asyncio